requirements.txt
amqp==5.1.1
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
billiard==4.1.0
blinker==1.6.2
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
connexion==2.14.2
cryptography==41.0.3
Flask==2.2.5
Flask-JWT-Extended==4.5.2
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kombu==5.3.1
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
packaging==23.1
passlib==1.7.4
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
PyYAML==6.0.1
redis==5.0.0
referencing==0.30.2
requests==2.31.0
rpds-py==0.9.2
six==1.16.0
SQLAlchemy==2.0.20
swagger-spec-validator==3.0.3
swagger-ui-bundle==0.0.9
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
zipp==3.16.2
